VitalStat Screen : Fixes the issue of not displaying body temp data in graph as the value was 5 or 6 decimal places like 36.944443 ->Round off the body temp value to two decimal places in order to show the data and change the data type of entry as double to plot double values in the graph

+++ app/src/main/java/com/shepherdapp/app/data/dto/add_vital_stats/vital_stats_dashboard/GraphData.kt
@@ -7,11 +7,11 @@ data class GraphData(
     @SerializedName("time")
     var day: String = "",
     @SerializedName("x")
-    var x: Int = 0,
+    var x: Double = 0.0,
     @SerializedName("x1")
-    var x1: Int = 0,
+    var x1: Double = 0.0,
     @SerializedName("y")
-    var y: Int = 0,
+    var y: Double = 0.0,
     @SerializedName("y2")
-    var y2: Int = 0
+    var y2: Double = 0.0
 )
